Current assignments box
