+++ src/scss/my-progress.scss
@@ -3,6 +3,17 @@
   width: 100%;
 }
 
+.text-flex-align {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+
+  &.height-40px {
+      height: 40px;
+  }
+}
+
 .auto-margin {
   margin: auto;
 }
